refactoring BuildDefinitionsList
